Added interface for parsing cmdline args
Basically just a bunch of functions/macros that make using <getopt.h> a bit more bearable. You register an Option, which has a short and long option (.ch and .longopt), and in doing so, the Option is added to the usage string in a formatted manner, is added to the long_opt[] array for getopt, and also the other one im forgetting. The 'class' is good enough for my purposes, but in future I may extract it and make it its own mini library.